+++ test/WhiteBoxTests.cpp
@@ -83,53 +83,53 @@ void WhiteBoxTests::testLOOLProtocolFunctions()
     CPPUNIT_ASSERT(LOOLProtocol::getTokenKeywordFromMessage(message, "mumble", map, mumble));
     CPPUNIT_ASSERT_EQUAL(2, mumble);
 
-    CPPUNIT_ASSERT_EQUAL(1UL, Util::trimmed("A").size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)Util::trimmed("A").size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), Util::trimmed("A"));
 
-    CPPUNIT_ASSERT_EQUAL(1UL, Util::trimmed(" X").size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)Util::trimmed(" X").size());
     CPPUNIT_ASSERT_EQUAL(std::string("X"), Util::trimmed(" X"));
 
-    CPPUNIT_ASSERT_EQUAL(1UL, Util::trimmed("Y ").size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)Util::trimmed("Y ").size());
     CPPUNIT_ASSERT_EQUAL(std::string("Y"), Util::trimmed("Y "));
 
-    CPPUNIT_ASSERT_EQUAL(1UL, Util::trimmed(" Z ").size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)Util::trimmed(" Z ").size());
     CPPUNIT_ASSERT_EQUAL(std::string("Z"), Util::trimmed(" Z "));
 
-    CPPUNIT_ASSERT_EQUAL(0UL, Util::trimmed(" ").size());
+    CPPUNIT_ASSERT_EQUAL(0UL, (unsigned long)Util::trimmed(" ").size());
     CPPUNIT_ASSERT_EQUAL(std::string(""), Util::trimmed(" "));
 
-    CPPUNIT_ASSERT_EQUAL(0UL, Util::trimmed("   ").size());
+    CPPUNIT_ASSERT_EQUAL(0UL, (unsigned long)Util::trimmed("   ").size());
     CPPUNIT_ASSERT_EQUAL(std::string(""), Util::trimmed("   "));
 
     std::string s;
 
     s = "A";
-    CPPUNIT_ASSERT_EQUAL(1UL, Util::trim(s).size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)Util::trim(s).size());
     s = "A";
     CPPUNIT_ASSERT_EQUAL(std::string("A"), Util::trim(s));
 
     s = " X";
-    CPPUNIT_ASSERT_EQUAL(1UL, Util::trim(s).size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)Util::trim(s).size());
     s = " X";
     CPPUNIT_ASSERT_EQUAL(std::string("X"), Util::trim(s));
 
     s = "Y ";
-    CPPUNIT_ASSERT_EQUAL(1UL, Util::trim(s).size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)Util::trim(s).size());
     s = "Y ";
     CPPUNIT_ASSERT_EQUAL(std::string("Y"), Util::trim(s));
 
     s = " Z ";
-    CPPUNIT_ASSERT_EQUAL(1UL, Util::trim(s).size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)Util::trim(s).size());
     s = " Z ";
     CPPUNIT_ASSERT_EQUAL(std::string("Z"), Util::trim(s));
 
     s = " ";
-    CPPUNIT_ASSERT_EQUAL(0UL, Util::trim(s).size());
+    CPPUNIT_ASSERT_EQUAL(0UL, (unsigned long)Util::trim(s).size());
     s = " ";
     CPPUNIT_ASSERT_EQUAL(std::string(""), Util::trim(s));
 
     s = "   ";
-    CPPUNIT_ASSERT_EQUAL(0UL, Util::trim(s).size());
+    CPPUNIT_ASSERT_EQUAL(0UL, (unsigned long)Util::trim(s).size());
     s = "   ";
     CPPUNIT_ASSERT_EQUAL(std::string(""), Util::trim(s));
 }
@@ -165,51 +165,51 @@ void WhiteBoxTests::testTokenizer()
     std::vector<std::string> tokens;
 
     tokens = LOOLProtocol::tokenize("");
-    CPPUNIT_ASSERT_EQUAL(0UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(0UL, (unsigned long)tokens.size());
 
     tokens = LOOLProtocol::tokenize("  ");
-    CPPUNIT_ASSERT_EQUAL(0UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(0UL, (unsigned long)tokens.size());
 
     tokens = LOOLProtocol::tokenize("A");
-    CPPUNIT_ASSERT_EQUAL(1UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), tokens[0]);
 
     tokens = LOOLProtocol::tokenize("  A");
-    CPPUNIT_ASSERT_EQUAL(1UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), tokens[0]);
 
     tokens = LOOLProtocol::tokenize("A  ");
-    CPPUNIT_ASSERT_EQUAL(1UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), tokens[0]);
 
     tokens = LOOLProtocol::tokenize(" A ");
-    CPPUNIT_ASSERT_EQUAL(1UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), tokens[0]);
 
     tokens = LOOLProtocol::tokenize(" A  Z ");
-    CPPUNIT_ASSERT_EQUAL(2UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(2UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), tokens[0]);
     CPPUNIT_ASSERT_EQUAL(std::string("Z"), tokens[1]);
 
     tokens = LOOLProtocol::tokenize("\n");
-    CPPUNIT_ASSERT_EQUAL(0UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(0UL, (unsigned long)tokens.size());
 
     tokens = LOOLProtocol::tokenize(" A  \nZ ");
-    CPPUNIT_ASSERT_EQUAL(1UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(1UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), tokens[0]);
 
     tokens = LOOLProtocol::tokenize(" A  Z\n ");
-    CPPUNIT_ASSERT_EQUAL(2UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(2UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), tokens[0]);
     CPPUNIT_ASSERT_EQUAL(std::string("Z"), tokens[1]);
 
     tokens = LOOLProtocol::tokenize(" A  Z  \n ");
-    CPPUNIT_ASSERT_EQUAL(2UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(2UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("A"), tokens[0]);
     CPPUNIT_ASSERT_EQUAL(std::string("Z"), tokens[1]);
 
     tokens = LOOLProtocol::tokenize("tile part=0 width=256 height=256 tileposx=0 tileposy=0 tilewidth=3840 tileheight=3840 ver=-1");
-    CPPUNIT_ASSERT_EQUAL(9UL, tokens.size());
+    CPPUNIT_ASSERT_EQUAL(9UL, (unsigned long)tokens.size());
     CPPUNIT_ASSERT_EQUAL(std::string("tile"), tokens[0]);
     CPPUNIT_ASSERT_EQUAL(std::string("part=0"), tokens[1]);
     CPPUNIT_ASSERT_EQUAL(std::string("width=256"), tokens[2]);
